Refactoring project
This moves all Packer configuration and provisioning scripts into a top-level ‘packer’ directory for better organisation.

(Note: At run-time a ‘packer_cache’ directory will be made at the top-level as well, but this is a temporary documentation and is ignored in the repository).

Because of the way I made improvements some provisioning files have changed during this refactoring. Namely:

- settings that Packer likes to control in VMware VMX files have been removed from the configuration template
- desktop builds use CentOS 7.8 minimal instead of 7.5 ‘full’
- cloud-init and related dependencies have been added to VMware builds
- the root account is now locked for security
- the system firewall is now enabled for security (SSH is allowed)
- wget removed as an installation dependency