Adding Terraform remote state
This is needed to provide an output for the latest desktop artefacts, similar to the primary image ID in DigitalOcean.

As we dont; have a provider project for the BAS VMware cluster (as we do for AWS and DO) this output needs to be tracked in this project. This could be changed in the future if we need a separate project.