Adding updated examples for provisioning desktop providers
This is needed to provide a reference for how people can use Terraform and cloud-init to create a compute resource in the BAS VMware cluster using a base image from this project.

As we don’t have a provider project for the BAS VMware cluster (as we do for AWS and DO) these examples need to be tracked in this project. This could be changed in the future if we need a separate project.